Added the handler call as a routine to not block
The handler was called from the main event loop meaning that it will be blocked for the duration of the handler execution. As long as this returns before the next tick it will still work, but we should not be relying on this. The task may take longer than the tick but the task might be handled by multiple workers etc